requirements.txt
# ============================================================
#  contrataoplanta.cl — requirements.txt
#  Python 3.11+
# ============================================================

# ── HTTP async y scraping ───────────────────────────────────
aiohttp==3.11.13
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# ── Base de datos ───────────────────────────────────────────
psycopg2-binary==2.9.9

# ── API FastAPI ─────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic[email]==2.6.3
slowapi==0.1.9

# ── Configuración ───────────────────────────────────────────
python-dotenv==1.0.1

# ── Utilidades de datos ─────────────────────────────────────
python-dateutil==2.9.0

# ── Emails transaccionales ──────────────────────────────────
resend==0.8.0

# ── Búsqueda rápida (Meilisearch) ──────────────────────────
meilisearch==0.31.6

# ── Scheduling opcional ─────────────────────────────────────
apscheduler==3.10.4

# ── Observabilidad ──────────────────────────────────────────
rich==13.7.1
